Rename to singlar to match discord docs
