+++ packages/types/src/discord/entitlement.ts
