+++ packages/types/src/index.ts
@@ -1,30 +1,30 @@
-export * from './discord/applications.js'
-export * from './discord/auditLogs.js'
+export * from './discord/application.js'
+export * from './discord/auditLog.js'
 export * from './discord/autoModeration.js'
-export * from './discord/channels.js'
-export * from './discord/emojis.js'
-export * from './discord/entitlements.js'
+export * from './discord/channel.js'
+export * from './discord/emoji.js'
+export * from './discord/entitlement.js'
 export * from './discord/gateway.js'
-export * from './discord/guilds.js'
-export * from './discord/guildScheduledEvents.js'
-export * from './discord/guildTemplates.js'
+export * from './discord/guild.js'
+export * from './discord/guildScheduledEvent.js'
+export * from './discord/guildTemplate.js'
 export * from './discord/interactions.js'
-export * from './discord/invites.js'
-export * from './discord/messages.js'
+export * from './discord/invite.js'
+export * from './discord/message.js'
 export * from './discord/oauth2.js'
 export * from './discord/opcodes.js'
 export * from './discord/permissions.js'
-export * from './discord/polls.js'
+export * from './discord/poll.js'
 export * from './discord/reference.js'
-export * from './discord/skus.js'
-export * from './discord/soundboards.js'
-export * from './discord/stageInstances.js'
-export * from './discord/stickers.js'
-export * from './discord/subscriptions.js'
-export * from './discord/users.js'
-export * from './discord/voice.js'
+export * from './discord/sku.js'
+export * from './discord/soundboard.js'
+export * from './discord/stageInstance.js'
+export * from './discord/sticker.js'
+export * from './discord/subscription.js'
 export * from './discord/teams.js'
+export * from './discord/user.js'
+export * from './discord/voice.js'
+export * from './discord/webhook.js'
 export * from './discord/webhookEvents.js'
-export * from './discord/webhooks.js'
 export * from './discordeno.js'
 export * from './shared.js'
